requirements.txt
llama-index
langchain
sentence-transformers
weaviate-client
pypdf
python-box
flask
python-multipart
streamlit